`SymbolTable` and `Scope` - working with `scope_id` represented as `usize`

+++ tytle_core/src/ast/semantic/scope.rs
@@ -1,7 +1,7 @@
 use crate::ast::semantic::{SymbolId, SymbolKind, Variable};
 use std::collections::HashMap;
 
-pub type ScopeId = u64;
+pub type ScopeId = usize;
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct Scope {
